Update Wunsch.md
Added "The description of Wunsch’s creative writing course for the 1936-1937 catalog cites that
the coursework was meant to develop in its students “A keenness of perception”; “Freshness of
expression”; and a “capacity for intellectual and emotional enjoyment.” The course focused on
the student’s development of the use of words, sense impressions, character analyses,
observations from daily life alongside observations from the individual imagination. These
perceptions would culminate in the latter part of the semester, when students would attempt
specific literary forms (with an emphasis on playwriting). While the course outcomes included
practice of writing in specific forms and genres, the material for student’s learning would be
gathered by their own perceptions from the “unique experience in…everyday life.”"